Refactor repository to own package
