+++ src/main/java/com/hedvig/notificationService/customerio/repository/InMemoryCustomerIOStateRepository.kt
@@ -1,8 +1,10 @@
-package com.hedvig.notificationService.customerio
+package com.hedvig.notificationService.customerio.repository
 
+import com.hedvig.notificationService.customerio.CustomerioState
 import java.time.Instant
 
-class InMemoryCustomerIOStateRepository(var data: Map<String, CustomerioState> = mapOf()) : CustomerIOStateRepository {
+class InMemoryCustomerIOStateRepository(var data: Map<String, CustomerioState> = mapOf()) :
+    CustomerIOStateRepository {
     override fun save(customerioState: CustomerioState) {
         data = data.plus(customerioState.memberId to customerioState)
     }
