+++ src/test/java/com/hedvig/notificationService/customerio/NorwaySignHackUpdateCustomerIOTest.kt
@@ -1,6 +1,7 @@
 package com.hedvig.notificationService.customerio
 
 import com.hedvig.customerio.CustomerioClient
+import com.hedvig.notificationService.customerio.repository.InMemoryCustomerIOStateRepository
 import io.mockk.MockKAnnotations
 import io.mockk.every
 import io.mockk.impl.annotations.MockK
@@ -30,7 +31,8 @@ class NorwaySignHackUpdateCustomerIOTest {
 
     lateinit var eventCreator: CustomerioEventCreator
 
-    private val repository = InMemoryCustomerIOStateRepository()
+    private val repository =
+        InMemoryCustomerIOStateRepository()
 
     @MockK
     lateinit var sut: CustomerioService
